proof of concept: library returns errors
motivation: #434

Proof of concept implementation of a mechanism to send error messages to users
of the wstunnel library.

Not an optimal API because it uses 3 different mechanisms for error
return.

The new signature is

`pub async fn run_client(args: Client, executor: impl TokioExecutor) -> anyhow::Result<Vec<anyhow::Result<()>>>`

Errors could be returned in the outer Result, in the inner Result, or in the new
error channel added to the LocalToRemote struct.

If we maintain it as an external fork we'll likely keep something like this to
keep changes minimal.

OTOH, we'd be happy to evolve this into something that can be merged.